feat: refactor Twitch authentication handling
